requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
google-cloud-storage==1.29.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
requests-futures
PyFunctional==1.2.0